Update java-jar.md (#233)
'wiremock' phrase leads to a misunderstanding. Because the usage -> 'wiremock' is not valid in command line. And in this case while wiremock server is up, user is getting no mapping found error when (s)he has made an http call. Perhaps user should know 'wiremock' usage is invalid in command line. But I have spent four hours to figure it out because I was tired. :)